Fix a few apparent TBA world consistency errors.
These are mostly cases of confusion between the "on the body"
and "about the body" slots (5 and 12).

But, there are also some that seem to just be misconfigured,
like non-weapons being wielded, worn items being unwearable,
weapons being worn instead of wielded, and items being worn
instead of held.

I avoided the cases where it was clear this was done intentionally,
like all the cases where the second neck was used for the missing
face slot (masks, etc).

The ones fixed here all seem to really be (minor) bugs.