TBA data file whitespace cleanups.
This just converts the few DOS text files still in here to standard text files (line-endings),
removes all the trailing spaces on lines, removes all the trailing blank lines, and replaces
all the tabs (except in .zon files, where they seem to be common) with expanded spaces.

It's easy to confirm this is actually a non-change, except for whitespace:
* `git show -w` shows this commit as only removing 8 trailing blank lines.

This should make no difference to tbaMUD itself, but it will make working on these files,
especially with scripts and automated tools, much easier.

This is the base change for a set of other changes I will put up once/if this merges,
without this, the others are just too complicated to wrangle.